Apply logit regression coefficients to logit post-standard LT qx values (and standard 45q15 and 5q0), along with input target (HIV-free, except for ZAF which is with-HIV) logit 45q15/5q0 to generate adjusted age-specific values

β1x and β2x are coefficients that vary by age x and which measure the impact of differences in child and adult mortality rates between a target life table and the standard life table on the estimated age pattern of mortality. In other words, both coefficients determine how much the estimated age pattern of mortality deviates from the standard by age and from linearity.

adjust_stan_qx(qx_stan, qx_target, modelpar_sims, id_vars)

Arguments

qx_stan
data.table with the variables ihme_loc_id, sex, year, sim, age, stan_qx, sq5, sq45. Results of calc_stan_lt to generate collapsed standard lifetables

qx_target
data.table with variables ihme_loc_id, year, sex, sim, target_45q15, target_5q0. Original target qx values, used as inputs into calc_stan_lt

modelpar_sims
data.table with variables sex, age, sim, sim_difflogit5, sim_difflogit45. Regression coefficients based on logit differences between standard LT and collapsed comparator LTs

id_vars
character vector of id variables (e.g. ihme_loc_id, sex, year, sim)

Value

returns the given data.table with mx for the 110 age-group (also modifies the given data.table in-place?)

Details

Betas created by taking country-time specific and region specific standard HIV-free lifetables, pairing with the region, and generating betas from the collapsed results See GBD2015 MortCoD appendix, page 37-38, for more details.