Calculate Standard Life Tables

Calculate a standard lifetable matched to a set of observed 5q0/45q15 observations (chunked by location/year/sex, parallel by draw) based off of the number of closest matches you want to keep in relation to each observation. Uses weights based on distance in geographic hierarchy from the empirical LT to the observation to weight the empirical collapse to the standard weight and collapse all empirical LT results into a single age-specific standard lifetable.

calc_stan_lt(lt_draw, empir_lt, weights, loc_map)

Arguments

lt_draw
numeric, refers to the sim number of hiv_free_qx to use (0-999).

empir_lt
data.table with variables: ihme_loc_id, year, sex, source_type, parent_ihme, super_region_id, region_id, empir_logit_45q15, empir_logit_5q0, lx* Empirical lifetable database

weights
data.table with variables: sex, age, region_cat (other/super region/region/country), weights Weights to determine relative weight of each matched database entry towards standard life table qx value

loc_map
data.table with variables: ihme_loc_id, region_name, super_region_name, parent_ihme

Value

list containing two data.tables: stan_lt: weighted standard qx values matching the input draws (the standard lifetable) variables: ihme_loc_id, year, sex, age, sim, qx, sq5, sq45 lt_list: each matched entry of lifetable results, with the normalized weight assigned to it variables: ihme_loc_id, year, sex, sim, stan_weight