Generate initial life tables from qx values of the first-stage iteration and regression parameters (HIV-free if non-ZAF, with-HIV if ZAF)

From post-iteration qx values, generate a full lifetable. Recalculate over-85 qx based on HMD parameters and age 80 qx estimates. Generate age 110 mx estimates based on age 105 mx estimates and regression parameters Recalculate under-10 nLx, mx, and ax based on Coale-Demeny child values

gen_stage_1_life_table(iter_qx, mx_110, betas_qx_85, ax_params)

Arguments

iter_qx
data.table with variables: sim, ihme_loc_id, year, sex, age (numeric), qx Post-stage-1 qx values

mx_110
data.table with variables: sex, parlnmx Parameter to multiply by ln 5m105 to obtain mx at age 110 (terminal age group)

betas_qx_85
data.table with variables: sex, age, beta_age, beta_logitqx80, constant Betas from logit model using selected high-quality HMD VR to convert from logit 5q80 to age-specific over-80 granular qx values. GBD2015 MortCoD Appendix: Part 3.2 in 3.3.A Note: age = seq(80,100,5)

ax_params
data.table with variables: sex, age, par_qx, par_sqx, par_con Constant and betas to generate ax values based on qx and qx^2 values. Unknown provenance

Value

data.table with variables: sex, age, qx, mx, ax, etc. Full lifetable with adjusted LT values from iterated qx

Examples