Import Empirical Lifetable values based on the lifetable database category for a given country

Import and format empirical lifetables based on the subset of lifetable entries which you want to import.

import_lt_empirical(lt_type = "", hiv_cdr, location)

Arguments

lt_type
character value representing the LT type to pull results for. Values: * large: Countries without country-specific lifetables, matching to general lifetable database. Will use all lifetables in the large lifetable database. * small: Countries with some country-specific non-generalizable lifetables. Will use all country-specific lifetables from the small database, and then all in large database. * USA: United States and all subnational locations. Will only use USA-specific lifetable. * ZAF: South Africa and all subnational locations. Will only use ZAF-specific lifetable. Note: not used as of GBD2016 -- switched to using with-HIV ZAF lifetables

hiv_cdr
data.table containing adult HIV crude death rate, to drop lifetables with high HIV results. Only used if lt_type == "large" variables: ihme_loc_id, sex, year, adult_hiv_cdr

Value

data.table with column names: * ihme_loc_id, year, sex, source_type * lx(0, 1, 5 (5) 110) * empir_5q0, empir_45q15, empir_logit_5q0, empir_logit_45q15 * region_id, super_region_id, parent_ihme

Examples