Wrapper to initialize variables for qx iteration process

Iterate scalar values to find ideal scalars to apply to age-specific qx or mx values, to better-match target 5q0 and 45q15 values Target values are generally HIV-free or with-HIV 5q0 and 45q15 calculated from entry simulations, with scaled HIV CDRs applied to them Initialize required variables for select_scalars function and the HIV adjustment functions called within. After scalars are selected, apply the best scalars to the final values to generate final age-specific adjusted values.

qx_iterations(current_lt, target_qx, hiv_type = "", scalar_set_max = 0.06,
  scalar_set_min = 1e-04, num_iterations = 30, num_scalars = 11, by_vars,
  age_group)

Arguments

current_lt
data.table with age-specific qx or mx values, along with aggregate 5q0 and 45q15 For stage 1, should contain standard LT qx; for stage 2, should contain mx values (if non-ZAF, HIV-free; if ZAF, with-HIV)

target_qx
data.table with calculated aggregate 5q0 and 45q15 qx values that iteration should attempt to match to.

hiv_type
character, options: "hiv_free", "with_hiv", "hiv_free_ZAF", "with_hiv_ZAF"

scalar_set_max
numeric, the initial upper value of scalars to be tested

scalar_set_min
numeric, the initial lower value of scalars to be tested

num_iterations
numeric whole number, the maximum number of times the function will test different ranges of scalars

num_scalars
numeric whole number, the number of scalar values to be tested within each iteration

by_vars
character vector, the variables within stan_qx and hiv_free_qx group by

age_group
character. Options: "5q0" or "45q15"

Value

a data table with by_vars, best_scalar, target_5q0, target_45q15. Includes qx if stage 1, mx and ax if stage 2.