Smooth post-standard LT matching lifetables for selected countries

To adjust for high-variance standard life table results over time, smooth the output qx values over years (but within id_vars such as age, country, sex, etc.)

smooth_stan_lts(stan_qx, id_vars)

Arguments

id_vars
character vector of id variables (should not include year), E.g. c("ihme_loc_id", "sex", "age", "sim")

dt
data.table with variables: ihme_loc_id, year, sex, age, sim, stan_qx

Value

data.table with variables: ihme_loc_id, year, sex, age, sim, stan_qx Contains smoothed qx values, should be same nrows as dt

Examples

test <- CJ(ihme_loc_id = "USA", year = c(1950:2016), sex = c("male", "female"), age = c(0, 1, seq(5, 5, 110)), sim = 1:3)

#> Error in eval(expr, envir, enclos): could not find function "CJ"

test[, qx := (age/110) + 1950/year-.5]

#> Error in eval(expr, envir, enclos): object 'test' not found

results <- smooth_stan_lts(test, id_vars = c("ihme_loc_id", "sex", "age", "sim"))

#> Error in smooth_stan_lts(test, id_vars = c("ihme_loc_id", "sex", "age",     "sim")): object 'test' not found